ref: app uses react router and move everything in homepage component
